Refactor to not gate FileHandle stuff on ProcessEnv
