+++ Sources/LanguageClient/DataChannel+LocalProcess.swift
@@ -5,26 +5,6 @@ import JSONRPC
 #if canImport(ProcessEnv)
 import ProcessEnv
 
-extension FileHandle {
-	public var dataStream: AsyncStream<Data> {
-		let (stream, continuation) = AsyncStream<Data>.makeStream()
-
-		readabilityHandler = { handle in
-			let data = handle.availableData
-
-			if data.isEmpty {
-				handle.readabilityHandler = nil
-				continuation.finish()
-				return
-			}
-
-			continuation.yield(data)
-		}
-
-		return stream
-	}
-}
-
 extension DataChannel {
 	@available(macOS 12.0, *)
 	public static func localProcessChannel(
